chatsaver/requirements.txt
llama-cpp-python>=0.3.0
transformers>=4.35.0
huggingface_hub>=0.19.0
sentencepiece>=0.2.0
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
requests>=2.30.0 